requirements.txt
pillow
orjson
docutils
pygments
pypiwin32
kivy_deps.sdl2==0.1.*
kivy_deps.glew==0.1.*
kivy_deps.angle==0.1.*
kivy==1.11.1